fixed multyscene syntax
