+++ js/editor-scene.js
@@ -4,11 +4,17 @@
 import URLParams from "./utils/url-params.js";
 import {XButton} from "./x-ui/x-button.js";
 import SceneMap from "./scene-map.js";
+import {ModeBasedScene} from "./scene-mode.js";
 import SceneModeEditor from "./scene-mode-editor.js";
 import SceneModePlayer from "./scene-mode-player.js";
 
-export default class PlatformerScene extends Phaser.Scene
+export default class EditorScene extends ModeBasedScene
 {
+  constructor()
+  {
+    super({key: "eritor", active: true});
+  }
+
   preload()
   {
     this.load.spritesheet(
@@ -27,10 +33,10 @@ export default class PlatformerScene extends Phaser.Scene
 
   create()
   {
-    this.isPaused = false;
-    this.isEditing = true;
-
     let map = this.map = new SceneMap(this, "map", "tilesSheet");
+    map.layers.forEach(
+      (l) => this.xui.insert(l.tilemapLayer)
+    );
 
     this.physics.world.setBounds(
       0, 0,
@@ -39,50 +45,48 @@ export default class PlatformerScene extends Phaser.Scene
       true, true, true, false
     );
 
-    this.editor = new SceneModeEditor(this, map.layer.tilemapLayer, map);
-    this.player = new SceneModePlayer(this, map.layer.tilemapLayer, map, map.editableTiles.coin);
-
-    this.player.on(SceneModePlayer.Events.playerWin, this.win, this);
-    this.player.on(SceneModePlayer.Events.playerDead, this.lose, this);
-
-    this.modes = {
-      editor: this.editor,
-      player: this.player
-    };
-
     this.params = URLParams.get();
-    this.setMode(
-      this.params.mode && this.params.mode == "player" ?
-        this.modes.player :
+    if (this.params.mode && this.params.mode == "player")
+    {
+      // PLayer Only
+      this.player = new SceneModePlayer(this, map.layer.tilemapLayer, map, map.editableTiles.coin);
+
+      this.player.on(SceneModePlayer.Events.playerWin, this.win, this);
+      this.player.on(SceneModePlayer.Events.playerDead, this.lose, this);
+
+      this.setMode(
+        this.player,
+        { levelId: this.params.levelId }
+      );
+      this.player.ui.first.buttons.replay.setVisible(false);
+    }
+    else
+    {
+      this.editor = new SceneModeEditor(this, map.layer.tilemapLayer, map);
+      this.player = new SceneModePlayer(this, map.layer.tilemapLayer, map, map.editableTiles.coin);
+
+      this.player.on(SceneModePlayer.Events.playerWin, this.win, this);
+      this.player.on(SceneModePlayer.Events.playerDead, this.lose, this);
+
+      this.modes = {
+        editor: this.editor,
+        player: this.player
+      };
+
+      this.setMode(
         this.modes.editor,
-      { levelId: this.params.levelId }
-    );
+        { levelId: this.params.levelId }
+      );
+
+      this.modes.editor.ui.first.buttons.play.on(
+        XButton.Events.onPointerUp, () => {this.mode = this.modes.player;});
+      this.modes.player.ui.first.buttons.replay.on(
+        XButton.Events.onPointerUp, () => {this.restart();});
+    }
     URLParams.clear();
 
-    this.modes.editor.ui.first.buttons.play.on(
-      XButton.Events.onPointerUp, () => {this.mode = this.modes.player});
-    this.modes.player.ui.first.buttons.replay.on(
-      XButton.Events.onPointerUp, () => {this.mode = this.modes.editor});
-  }
-
-  get mode()
-  {
-    return this.__mode;
-  }
-
-  setMode(newMode, params)
-  {
-    if (this.__mode == newMode) return;
-    if (this.__mode)
-      this.__mode.leave();
-
-    this.__mode = newMode;
-    this.__mode.enter(params);
-  }
-
-  set mode(newMode)
-  {
-    this.setMode(newMode)
+    this.isPaused = false;
+    console.log(this);
   }
 
   update(time, delta) {
@@ -93,11 +97,11 @@ export default class PlatformerScene extends Phaser.Scene
   win()
   {
     this.isPaused = true;
-    let cam = this.cameras.main;
 
     this.player.player.freeze();
+
     this.xui.showDialog("Вы успешно создали уровень!\nХотите сохранить его?",
-                          ["replay", "cancel", "ok"])
+                        ["replay", "cancel", "ok"])
     .then(
       // OK
       () => this.map.level.push().then(
@@ -129,8 +133,7 @@ export default class PlatformerScene extends Phaser.Scene
 
   lose()
   {
-    this.isPaused = true;
-
+    this.player.player.freeze();
     const cam = this.cameras.main;
     cam.shake(100, 0.05);
 
@@ -140,20 +143,18 @@ export default class PlatformerScene extends Phaser.Scene
   restart(clearLevel)
   {
     this.isPaused = true;
+
     const cam = this.cameras.main;
     cam.fade(250, 0, 0, 0);
 
     if (clearLevel)
     {
-      this.urlParams = {};
       localStorage.clear();
     }
 
-    this.player.player.freeze();
-
     cam.once("camerafadeoutcomplete", () => {
-      this.editor.destroy();
-      this.player.destroy();
+      if (this.editor) this.editor.destroy();
+      if (this.player) this.player.destroy();
       this.__mode = undefined;
       this.scene.restart();
     });
